requirements.txt
Flask==3.0.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.2
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
werkzeug==3.0.1

# For sentiment analysis (optional but recommended)
textblob==0.17.1

# Note: After installing requirements, run:
# python -m spacy download en_core_web_sm